+++ pkg/common/constants/constants.go
@@ -61,6 +61,7 @@ const (
 	MpsAnnotation                 = "mps"
 	StalePodgroupTimeStamp        = "kai.scheduler/stale-podgroup-timestamp"
 	LastStartTimeStamp            = "kai.scheduler/last-start-timestamp"
+	LastEvictionTimeStamp         = "kai.scheduler/last-eviction-timestamp"
 	GpuSharingConfigMapAnnotation = "runai/shared-gpu-configmap"
 	NvidiaVisibleDevices          = "NVIDIA_VISIBLE_DEVICES"
 	NumaPlacementPredicted        = "kai.scheduler/numa-placement-predicted"