+++ pkg/podgrouper/podgrouper/plugins/constants/constants.go
@@ -4,11 +4,12 @@
 package constants
 
 const (
-	PodGroupNamePrefix     = "pg"
-	ProjectLabelKey        = "project"
-	PriorityLabelKey       = "priorityClassName"
-	PreemptibilityLabelKey = "kai.scheduler/preemptibility"
-	UserLabelKey           = "user"
+	PodGroupNamePrefix      = "pg"
+	ProjectLabelKey         = "project"
+	PriorityLabelKey        = "priorityClassName"
+	PreemptibilityLabelKey  = "kai.scheduler/preemptibility"
+	PreemptionDelayLabelKey = "kai.scheduler/preemption-delay"
+	UserLabelKey            = "user"
 
 	BuildPriorityClass     = "build"
 	TrainPriorityClass     = "train"